CameraFeed/requirements.txt
kafka-python>=2.0.0,<3.0.0
python-dateutil
requests
shapely
PyYAML
